Island: Move Role and User in authentication_service dir
