+++ monkey/monkey_island/cc/app.py
@@ -12,7 +12,6 @@ from wtforms import StringField, ValidationError
 
 from common import AccountRole, DIContainer
 from monkey_island.cc.flask_utils import FlaskDIWrapper
-from monkey_island.cc.models import Role, User
 from monkey_island.cc.resources import (
     AgentBinaries,
     AgentEvents,
@@ -43,6 +42,8 @@ from monkey_island.cc.resources.version import Version
 from monkey_island.cc.server_utils import generate_flask_security_configuration
 from monkey_island.cc.server_utils.consts import MONKEY_ISLAND_ABS_PATH
 from monkey_island.cc.services import register_agent_configuration_resources
+from monkey_island.cc.services.authentication_service.role import Role
+from monkey_island.cc.services.authentication_service.user import User
 from monkey_island.cc.services.representations import output_json
 from monkey_island.cc.setup.mongo.mongo_setup import MONGO_DB_HOST, MONGO_DB_NAME, MONGO_DB_PORT
