+++ monkey/monkey_island/cc/services/authentication_service/authentication_service.py
@@ -1,8 +1,9 @@
 from pathlib import Path
 
 from monkey_island.cc.event_queue import IIslandEventQueue, IslandEventTopic
-from monkey_island.cc.models import IslandMode, User
+from monkey_island.cc.models import IslandMode
 from monkey_island.cc.server_utils.encryption import ILockableEncryptor
+from monkey_island.cc.services.authentication_service.user import User
 
 
 class AuthenticationService:
