+++ monkey/monkey_island/cc/services/authentication_service/user.py
